// frontend/src/shared/api/generated/model/riskAlert.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * StockQuest API
 * Stock Trading Simulation Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { RiskAlertAlertType } from "./riskAlertAlertType";
import type { RiskAlertSeverity } from "./riskAlertSeverity";
import type { RiskAlertStatus } from "./riskAlertStatus";

export interface RiskAlert {
  /** 알림 유형 */
  alertType?: RiskAlertAlertType;
  /** 현재 값 */
  currentValue?: string;
  /** 알림 ID */
  id?: number;
  /** 알림 메시지 */
  message?: string;
  /** 포트폴리오 ID */
  portfolioId?: number;
  /** 해결 시간 */
  resolvedAt?: string;
  /** 심각도 */
  severity?: RiskAlertSeverity;
  /** 알림 상태 */
  status?: RiskAlertStatus;
  /** 임계값 */
  thresholdValue?: string;
  /** 발생 시간 */
  triggeredAt?: string;
}